refactor: move time utils to common
