+++ AnkiDroid/src/main/java/com/ichi2/anki/BackupManager.kt
@@ -17,11 +17,11 @@
 package com.ichi2.anki
 
 import android.text.format.DateFormat
+import com.ichi2.anki.common.time.Time
+import com.ichi2.anki.common.time.Time.Companion.utcOffset
+import com.ichi2.anki.common.time.TimeManager
 import com.ichi2.libanki.Collection
 import com.ichi2.libanki.Utils
-import com.ichi2.libanki.utils.Time
-import com.ichi2.libanki.utils.Time.Companion.utcOffset
-import com.ichi2.libanki.utils.TimeManager
 import com.ichi2.utils.FileUtil.getFreeDiskSpace
 import timber.log.Timber
 import java.io.File
