+++ AnkiDroid/src/main/java/com/ichi2/anki/dialogs/BackupPromptDialog.kt
@@ -25,6 +25,7 @@ import com.ichi2.anki.CollectionManager.withCol
 import com.ichi2.anki.CrashReportService
 import com.ichi2.anki.DeckPicker
 import com.ichi2.anki.R
+import com.ichi2.anki.common.time.TimeManager
 import com.ichi2.anki.isLoggedIn
 import com.ichi2.anki.millisecondsSinceLastSync
 import com.ichi2.anki.preferences.sharedPrefs
@@ -32,7 +33,6 @@ import com.ichi2.anki.servicelayer.ScopedStorageService.collectionWillBeMadeInac
 import com.ichi2.anki.servicelayer.ScopedStorageService.userIsPromptedToDeleteCollectionOnUninstall
 import com.ichi2.compat.CompatHelper.Companion.getPackageInfoCompat
 import com.ichi2.compat.PackageInfoFlagsCompat
-import com.ichi2.libanki.utils.TimeManager
 import com.ichi2.utils.Permissions
 import com.ichi2.utils.cancelable
 import com.ichi2.utils.checkBoxPrompt
