+++ AnkiDroid/src/main/java/com/ichi2/libanki/Storage.kt
@@ -15,9 +15,9 @@
  */
 package com.ichi2.libanki
 
+import com.ichi2.anki.common.time.Time
+import com.ichi2.anki.common.time.TimeManager.time
 import com.ichi2.anki.getDayStart
-import com.ichi2.libanki.utils.Time
-import com.ichi2.libanki.utils.TimeManager.time
 import net.ankiweb.rsdroid.Backend
 import net.ankiweb.rsdroid.BackendFactory
 import java.io.File
