+++ AnkiDroid/src/test/java/com/ichi2/anki/RobolectricTest.kt
@@ -33,6 +33,7 @@ import androidx.work.Configuration
 import androidx.work.testing.SynchronousExecutor
 import androidx.work.testing.WorkManagerTestInitHelper
 import com.ichi2.anki.CollectionManager.CollectionOpenFailure
+import com.ichi2.anki.common.time.TimeManager
 import com.ichi2.anki.dialogs.DialogHandler
 import com.ichi2.anki.preferences.sharedPrefs
 import com.ichi2.compat.customtabs.CustomTabActivityHelper
@@ -41,7 +42,6 @@ import com.ichi2.libanki.ChangeManager
 import com.ichi2.libanki.Collection
 import com.ichi2.libanki.NotetypeJson
 import com.ichi2.libanki.Storage
-import com.ichi2.libanki.utils.TimeManager
 import com.ichi2.testutils.AndroidTest
 import com.ichi2.testutils.MockTime
 import com.ichi2.testutils.TaskSchedulerRule
