+++ AnkiDroid/src/test/java/com/ichi2/libanki/SchedulerTest.kt
@@ -18,6 +18,8 @@ package com.ichi2.libanki
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import anki.scheduler.UnburyDeckRequest
 import com.ichi2.anki.Ease
+import com.ichi2.anki.common.time.TimeManager
+import com.ichi2.anki.common.time.TimeManager.time
 import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import com.ichi2.anki.utils.SECONDS_PER_DAY
 import com.ichi2.libanki.CardType.Relearning
@@ -25,8 +27,6 @@ import com.ichi2.libanki.Consts.LEECH_SUSPEND
 import com.ichi2.libanki.Consts.STARTING_FACTOR
 import com.ichi2.libanki.exception.ConfirmModSchemaException
 import com.ichi2.libanki.sched.Counts
-import com.ichi2.libanki.utils.TimeManager
-import com.ichi2.libanki.utils.TimeManager.time
 import com.ichi2.testutils.AnkiAssert
 import com.ichi2.testutils.JvmTest
 import com.ichi2.testutils.ext.addNote
