+++ common/src/main/java/com/ichi2/anki/common/time/SystemTime.kt
@@ -13,7 +13,9 @@
  You should have received a copy of the GNU General Public License along with
  this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.ichi2.libanki.utils
+package com.ichi2.anki.common.time
+
+import com.ichi2.anki.common.time.Time
 
 class SystemTime : Time() {
     override fun intTimeMS(): Long = System.currentTimeMillis()
